Removed auto layer coallocation determiniation
Previously, if  was not provided to KFAC, KFAC would determine if the A and G factors for an individual layer should be computed on the same worker or different workers depending on the number of available workers. This is now removed and by default  to reduce any confusion on where factors are being placed.


Former-commit-id: befb94be1805f5c322088f2f31645c5bc3720853