Added LSTM cell variants
Usually the weight matrices for the gates in an LSTM cell are combined into one larger matrix.
For KFAC, however, this produces larger factors that need to be inverted so it may be preferable to keep the gates as separate Linear modules. This commit adds LSTMCell classes with both implementations.
The LSTMCell is faster than LSTMCellKFAC for both KFAC and SGD however when you compare the per iteration time between KFAC and SGD for both types, KFAC has less of an overhead compared to SGD when using LSTMCellKFAC because the factors are smaller and easier to invert.
In the future, the KFACLayer class that wraps the Linear modules in LSTMCell should automatically split the weights into 4 parts, invert, and then apply so that we get the advantage of faster forward/backward passes by keeping one large weight matrix but we save on inversion time.


Former-commit-id: 71f2e6f244acc3bdf48b8e3c775d95a4fcd78419