Added KFAC communication method variants
Added a new parameter, comm_method, to KFAC(). comm_method is of type kfac.CommMethod, an enum defining the two communication protocols that KFAC supports. The two methods are COMM_OPT and MEM_OPT which optimize for reduced communication and reduced memory respectively. These variants are further explained and investigated in the paper, where COMM_OPT=KFAC_opt and MEM_OPT=KFAC_lw. The TL;DR is that with COMM_OPT, inverses are computed indepently from the preconditioned gradient allowing for reduced communication in non-KFAC update steps. In MEM_OPT, the inverses and preconditioned gradients are computed for a single layer on one worker such that the inverses only need to be stored on one worker (instead of duplicates on every worker) but this requires additional communication in every step regardless of if the inverses are updated.


Former-commit-id: 7fe93e02cd6f8388e58043ba1018daa6a5edde9f