Added flag to store factors in fp16
Added store_factors_as_fp16 parameters to KFAC which will designate the KFACLayers to store factors in fp16 instead of the default fp32. This will reduce the memory consumption of KFAC and slightly improve performance if your device supports fp16 operations. Note that the factor will be cast to fp32 before inversion/decomposition because those operations are not stable in fp16. It is possible that the loss of precision could hurt overall performance so this flag should be used with care


Former-commit-id: 892985f8adeb90fdd4b7a4e9b0f5148df2047ee9