FIX: MEM_OPT bug from previous commit
The previous commit broke MEM_OPT because the comm.backend was made global so no methods needed to accept a rank as input.
Thus, KFACLayer.compute_preconditioned_gradient() no longer accepted rank as an argument and could not check if only a single rank was going to compute the gradient so all ranks ended up computing the gradient.
This commit adds the compute_grad_ranks attribute to KFACLayer which contains a list of all ranks that shoudl execute compute_preconditioned_gradient().
In MEM_OPT, compute_grad_ranks has size 1, and in COMM_OPT, compute_grad_ranks is a list of all ranks because the ranks compute the gradients in parallel locally.


Former-commit-id: bc6d5a3a80e13445d036dff3fd8a4c7f8da8d7b0