ADD: optimized gradient preconditioning
Commit 78473b7 remove the QQT optimization for gradient preconditioning, and this commit adds back a more correct implementation.
Specifically, the prediv_eigenvalues flag in the KFACLayer class will, when True, compute 1 / (dG * dA.T + damping) once on a single inverse worker and broadcast the results to the remaining gradient worker ranks.
The benefit is that dA and dG do not change during non-K-FAC update steps so in the old code, 1 / (dG * dA.T + damping) was computed redundantly every iteration on every worker.
This uses more memory as dG * dA.T is a matrix whereas dG and dA are vectors; however, the gradient preconditioning time on Cifar10 was reduced by 22%.
Note that using the prediv_eigenvalues flag requires the A and G inverse ranks to be the same for each layer.

Other changes:
 - default coallocate_layer_factors=True in examples
 - refactored KFACLayer to store data in self.state, a dictionary that holds the factors, inverses, or eigendecompositions
 - KFAC.memory_usage() now calculates any tensors in the KFACLayer.state and the preconditioned gradients.


Former-commit-id: 374cbf82142a97225f16d77d2c7a394328225a48