Fix in-place operation error with factor accumulation
When accumulating the inputs for the A factor, PyTorch was raising an error that variables needed for the gradient computation were modified by in-place operations.
This was because on the first accumulation step, we were saving a reference to the intermediate data.
In this fix, we simply save a reference to the data, rather than the tensor as only the tensor has a history.


Former-commit-id: ed4ec4e335a1915e9190f299a6bdf13856483cbb