Improve mini-step tracking for factor accumulation
The mini-step tracking assumed that forward/backward steps were never overlapped so a single counter was used to determine when a gradient accumulation boundary was.
This assumption breaks for more advanced layer pipelining schemes such as those used by DeepSpeed.
The fix is counting the mini-steps on a per-module basis.


Former-commit-id: 4ebd7bb43db3f19b3975db1f77fa1d7a8a1fb968