Fix BaseKFACPreconditioner WorkAssignment bug
Previously, the BaseKFACPreconditioner was not correctly following the
WorkAssignment strategy essentially resulting in every rank computing
the inverse for the layer. As a result, the unit tests did not catch
subsequent bugs with the broadcast/preconditioning strategy not being
followed either.

This commit fixes the above issues so there is no longer redundant
computation and the WorkAssignment provided to BaseKFACPreconditioner is
followed.