Add groups for factor allreduce
In KAISA, every rank participates in the factor allreduce so specifying
the group was not necessary. However, future support for model/pipeline-
parallel methods will mean that not every rank participates in the
factor allreduce for every layer so this change adds functionality for
the WorkAssignment to specify the communication group to use.