Add state dict save/load for GPT
Adds two methods of saving/loading the state dict: the standard way
where calling state_dict() returns a single dict with all of the factors
for every layer and an alternative method of specifying a checkpoint
dir and each inverse worker saves the state dict for its layers to a
unique file in the state dict.

This new method was added because for larger GPT models, its is not
practical to create a single state dict with all the data and gather
it onto a single worker.