Fix layer registration regression (#53)
Previously, specifying a parent module type as a skip_layer would cause
the KFAC registration system to not recurse into modules of that type.
This changed in v0.4 because the layers in the model are flattened
before registration.

To fix this regression, skip_layers is now a list of regex patterns that
get applied to a modules name and class name. Since module names encode
the inheritance of layers, a regex can be specified that will prevent
recursion into certain parts of a module.